Add httpie CLI tool
This python based tool is supposed to be better/easier than curl, so
figured we'd give it a shot.

Hattip to makandropedia: https://makandracards.com/makandra/47945-httpie-nice-command-line-http-client

Explain the problem that this commit is solving. Why, not how.
Are there side effects or other unintuitive consequenses of this
change? Here's the place to explain them.

Resolves: #123
See also: #456, #789