Add shell script to toggle pg11 and latest
As long as two or more versions of Postgresql are running, there's going
to be a need to switch between them. The `.envrc` file provided by `direnv`
can do this, but its trickier than it seems since there's no clear way to
test whether a single service is running from homebrew.

This uses `-c` to get a *line count* of matching lines, which thanks to the
pair of `grep` commands will either be 1 or 0, so we can set up the
appropriate `if` commands in the bash script.